Made sign up and otp screens responsive
